+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Replace checkout line items amount `code` field with encrypted amount field.
 - Rename checkout lines items `id` field to `product_id`.
+- Refactor checkout line items amount functions.
 
 ### Removed
 - Checkout line items data validation.